Trigger webhook setup 1 second after server start
This project was constantly failing to create webhooks. It turns out the
issue was that, in order to create a webhook, the server had to respond
to the HEAD request being made by Trello. So now the initial webhook
requests are triggered after the server starts, so long as the server
starts in less than a second.

I'd love a cleaner way to do this, but it works for now.